Use getentropy() if possible
This uses the getentropy() wrapper for getrandom(2) if possible. This
was introduced in glibc 2.25. Different libc implementations already
provide it. If the call fails for whatever reason (e.g., new glibc,
outdated kernel), the code falls back to reading from /dev/urandom.

+++ user/shared/shared_tool.h
@@ -88,7 +88,7 @@ extern int lk_bdev_delete(const unsigned minor);
  * caller should free(bd->bd_name) once it is no longer needed. */
 extern int lk_bdev_load(const unsigned minor, struct bdev_info *bd);
 
-extern void get_random_bytes(void* buffer, int len);
+extern void get_random_bytes(void *buffer, size_t len);
 
 /* Since glibc 2.8~20080505-0ubuntu7 asprintf() is declared with the
    warn_unused_result attribute.... */
